feat: Replaced modal into a new create-event page
